services/recuperar_embeddings/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.35.3
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
